Update NIFU_Synthesis.py
Changed button to apply changes to pump to say "Apply to Pumps", allow for the user to stay on the commands page after issuing a command, and added an extra column to the pump commands for the balance port numbers. Also, made the checkboxes for what to plot to appear after checking what data is wanted to be graphed.